refactor(events): reuse FieldError in attendee-input-field
Move FieldError to its own module (avoiding a circular import with
event-form-sections, which imports AttendeeInputField) and use it in
attendee-input-field instead of a hand-rolled error paragraph, so error
styling lives in one place.
